Added runConfigurations for PyCharm users.  One change to support GitHub pages supporting html pages in docs for the project.  Another change for building various themes in Sphinx for help in explaining the usage of Sphinx, and it's importance.
Add Argument parsing to __main__.py.  Intend to show usage of built
in argument parser, how help is created from it, and call it out
as a best practice.

Setup.py added as best practice.

Updates to README

Updates to requirements.txt for Sphinx themes in use by the project.