#9 Add the project.cfg infrastructure. This is done to store version information, build information, and other data points that are used across multiple files (for instance conf.py with sphinx and setup.py with the build process).
Comment in Makefile to remember what $@ syntax meant in the bash shell.

Update .gitignore with exclude for venv because it's not catching the
exclusion with the other patterns in there.

dist_pypi.bat updated with the same updates done in the browser_driver
project.

Bumped version number and updated Readme.  Updated requirements.txt
for updated versions necessary for updates in other dependencies found
while building docs.

setup.py updated to use the project.cfg file for version and build
information